Moved REST tests out of subdir
